edit entries feature fixed

+++ vault/storeold.php
@@ -9,7 +9,7 @@ $conn = mysqli_connect('db', 'root', 'MYSQL_ROOT_PASSWORD', 'PM_1');
 if ($conn->connect_error) {
     die("Connection failed: " . $conn->connect_error);
 }
-$sql1 = "SELECT * FROM `Old_Passwords` WHERE `User_ID`='$userid'";
+$sql1 = "SELECT * FROM `Old_Passwords` WHERE `User_ID`='$userid' AND `Website`='" . $_SESSION['website'] . "'";
 $res1 = mysqli_query($conn, $sql1);
 if (!$res1) {
     header("Refresh:3, url=/vault");
@@ -18,7 +18,7 @@ if (!$res1) {
     exit();
 }
 if ($res1->num_rows === 5) {
-    $sql2 = "DELETE FROM `Old_Passwords` WHERE `Add_Date` IS NOT NULL ORDER BY `Add_Date` DESC LIMIT 1";
+    $sql2 = "DELETE FROM `Old_Passwords` WHERE `User_ID`='$userid' AND `Website`='" . $_SESSION['website'] . "' AND `Add_Date` IS NOT NULL ORDER BY `Add_Date` DESC LIMIT 1";
     $res2 = mysqli_query($conn, $sql2);
     if (!$res2) {
         header("Refresh:3, url=/vault");
@@ -31,7 +31,7 @@ $link = $_SESSION['link'];
 $hash = $_SESSION['hash'];
 $website = $_SESSION['website'];
 $adddate = $_SESSION['adddate'];
-$sql3 = "INSERT INTO `Old_Passwords` VALUES ('$userid','$website','$link','$hash','$adddate')";
+$sql3 = "INSERT INTO `Old_Passwords`(`User_ID`,`Website`,`Link`,`Old_Hash`,`Add_Date`) VALUES ('$userid','$website','$link','$hash','$adddate')";
 $res3 = mysqli_query($conn, $sql3);
 $conn->close();
 if (!$res3) {
